move get_signers_from_chain
