+++ crates/client/src/user.rs
@@ -18,7 +18,6 @@ use crate::{
     substrate::query_chain,
 };
 use entropy_shared::{user::ValidatorInfo, BlockNumber, HashingAlgorithm};
-use rand::prelude::SliceRandom;
 use serde::{Deserialize, Serialize};
 use subxt::{backend::legacy::LegacyRpcMethods, OnlineClient};
 
@@ -54,67 +53,6 @@ pub struct RelayerSignatureRequest {
     pub signature_verifying_key: Vec<u8>,
     /// Information for the validators in the signing party
     pub validators_info: Vec<ValidatorInfo>,
-}
-
-/// Returns a threshold of signer's ValidatorInfo from the chain
-pub async fn get_signers_from_chain(
-    api: &OnlineClient<EntropyConfig>,
-    rpc: &LegacyRpcMethods<EntropyConfig>,
-) -> Result<Vec<ValidatorInfo>, SubgroupGetError> {
-    let signer_query = entropy::storage().staking_extension().signers();
-    let signers = query_chain(api, rpc, signer_query, None)
-        .await?
-        .ok_or_else(|| SubgroupGetError::ChainFetch("Get all validators error"))?;
-
-    let key_info_query = entropy::storage().parameters().signers_info();
-    let threshold = query_chain(api, rpc, key_info_query, None)
-        .await?
-        .ok_or_else(|| SubgroupGetError::ChainFetch("Failed to get signers info"))?
-        .threshold;
-
-    let selected_signers: Vec<_> = {
-        let mut cloned_signers = signers.clone();
-        // TODO: temp remove dave for now until test dave is spun up correctly
-        cloned_signers.pop();
-        cloned_signers
-            .choose_multiple(&mut rand::thread_rng(), threshold as usize)
-            .cloned()
-            .collect()
-    };
-
-    let block_hash = rpc.chain_get_block_hash(None).await?;
-    let mut handles = Vec::new();
-
-    for signer in selected_signers {
-        let handle: tokio::task::JoinHandle<Result<ValidatorInfo, SubgroupGetError>> =
-            tokio::task::spawn({
-                let api = api.clone();
-                let rpc = rpc.clone();
-                async move {
-                    let threshold_address_query =
-                        entropy::storage().staking_extension().threshold_servers(signer);
-                    let server_info = query_chain(&api, &rpc, threshold_address_query, block_hash)
-                        .await?
-                        .ok_or_else(|| {
-                            SubgroupGetError::ChainFetch("threshold_servers query error")
-                        })?;
-                    Ok(ValidatorInfo {
-                        x25519_public_key: server_info.x25519_public_key,
-                        ip_address: std::str::from_utf8(&server_info.endpoint)?.to_string(),
-                        tss_account: server_info.tss_account,
-                    })
-                }
-            });
-
-        handles.push(handle);
-    }
-
-    let mut all_signers: Vec<ValidatorInfo> = vec![];
-    for handle in handles {
-        all_signers.push(handle.await??);
-    }
-
-    Ok(all_signers)
 }
 
 /// Gets a validator from chain to relay a message to the signers
